Extract JWKS/registry discovery functions to utils/registry_discovery
Move GetNSIssuerURL, GetJWKSURLFromIssuerURL, and GetJWKSFromIssUrl
from server_utils/registry.go into a new utils/registry_discovery
package and delete the original file. This breaks the dependency
chain that previously forced client-only code paths to import
server_utils and its transitive server dependencies.

Note that I couldn't stick these directly into a utils file because
utils can't import config (and these require federation discovery).
I opted to keep them colocated with utils but in their own package
to increase the likelihood someone finds them when they need these
types of functions.
