+++ director/origin_api.go
@@ -37,9 +37,9 @@ import (
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pelicanplatform/pelican/server_structs"
-	"github.com/pelicanplatform/pelican/server_utils"
 	"github.com/pelicanplatform/pelican/token_scopes"
 	"github.com/pelicanplatform/pelican/utils"
+	"github.com/pelicanplatform/pelican/utils/registry_discovery"
 )
 
 var (
@@ -107,12 +107,12 @@ func checkNamespaceStatus(prefix string, registryWebUrlStr string) (bool, error)
 // see if the entity is authorized to advertise an origin for the
 // namespace
 func verifyAdvertiseToken(ctx context.Context, token, namespace string) (bool, error) {
-	issuerUrl, err := server_utils.GetNSIssuerURL(namespace)
+	issuerUrl, err := registry_discovery.GetNSIssuerURL(namespace)
 	if err != nil {
 		return false, errors.Wrap(err, "failed to get issuer for namespace "+namespace)
 	}
 
-	keyLoc, err := server_utils.GetJWKSURLFromIssuerURL(issuerUrl)
+	keyLoc, err := registry_discovery.GetJWKSURLFromIssuerURL(issuerUrl)
 	if err != nil {
 		return false, errors.Wrap(err, "failed to get JWKS URL from the issuer URL at "+issuerUrl)
 	}
