+++ utils/registry_discovery/registry_discovery.go
@@ -16,7 +16,7 @@
  *
  ***************************************************************/
 
-package server_utils
+package registry_discovery
 
 import (
 	"context"
